requirements.txt
matplotlib==3.5.3 # version required for plotnine
plotnine==0.10.1 # Python version of ggplot2
numpy>=1.22
pandas==1.4.2
scikit-learn==1.1.3
sqlalchemy==1.4.32
joblib==1.2.0
sqlite==3.40.0
python-dotenv==0.21.0
pymysql==1.0.2
cryptography==38.0.3
mysql-connector==2.2.9
### UTILS
jupyterlab==3.4.2
tqdm==4.62.0